requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
extra-streamlit-components==0.1.80
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.10
lxml==5.3.0
MarkupSafe==3.0.2
narwhals==1.34.1
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
plotly==6.0.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.44.1
streamlit-authenticator==0.4.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
matplotlib
streamlit
sqlite3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
